POP refactoring for Container CBZ support wip
